refactor(132): delete CatalogOverview, inline map into StacBrowser
CatalogOverview was a combined map+timeline component. StacBrowser now
provides the timeline in a dedicated GoldenLayout panel, making the
CatalogOverview wrapper redundant.

Changes:
- Inline react-leaflet map directly in StacBrowser's map panel renderer
  (MapContainer, Rectangle, TileLayer, FitBounds, ViewportTracker)
- Remove CatalogOverview component export from index.ts
- Delete entire CatalogOverview directory (component, CSS, tests, stories)
- CatalogOverviewItem type still exported from filter-engine/types.ts
