--- shared/components/src/CatalogOverview/CatalogOverview.css
@@ -1,144 +1,0 @@
-.catalog-overview {
-  --co-bg: var(--vscode-editor-background, #1e1e1e);
-  --co-fg: var(--vscode-editor-foreground, #cccccc);
-  --co-accent: var(--vscode-focusBorder, #007fd4);
-  --co-bar-fill: var(--vscode-charts-blue, #3794ff);
-  --co-bar-hover: var(--vscode-charts-yellow, #cca700);
-  --co-border: var(--vscode-panel-border, #444444);
-  --co-tooltip-bg: var(--vscode-editorHoverWidget-background, #252526);
-  --co-tooltip-fg: var(--vscode-editorHoverWidget-foreground, #cccccc);
-  --co-dragbar: var(--vscode-sash-hoverBorder, #007fd4);
-
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  background: var(--co-bg);
-  color: var(--co-fg);
-  overflow: hidden;
-  font-family: var(--vscode-font-family, system-ui, sans-serif);
-  font-size: var(--vscode-font-size, 13px);
-}
-
-/* Map region */
-.catalog-overview__map {
-  flex: var(--co-map-flex, 6);
-  min-height: 80px;
-  position: relative;
-}
-
-.catalog-overview__map .leaflet-container {
-  width: 100%;
-  height: 100%;
-  background: var(--co-bg);
-}
-
-/* Drag bar */
-.catalog-overview__dragbar {
-  height: 6px;
-  background: var(--co-border);
-  cursor: row-resize;
-  flex-shrink: 0;
-  transition: background 0.15s;
-}
-
-.catalog-overview__dragbar:hover,
-.catalog-overview__dragbar--active {
-  background: var(--co-dragbar);
-}
-
-/* Timeline region */
-.catalog-overview__timeline {
-  flex: var(--co-timeline-flex, 4);
-  min-height: 60px;
-  overflow-y: auto;
-  overflow-x: hidden;
-  padding: 8px;
-}
-
-.catalog-overview__timeline svg {
-  display: block;
-  width: 100%;
-}
-
-/* Timeline elements */
-.catalog-overview__timeline-bar {
-  fill: var(--co-bar-fill);
-  cursor: pointer;
-  transition: fill 0.15s;
-}
-
-.catalog-overview__timeline-bar:hover {
-  fill: var(--co-bar-hover);
-}
-
-.catalog-overview__timeline-point {
-  fill: var(--co-bar-fill);
-  cursor: pointer;
-}
-
-.catalog-overview__timeline-point:hover {
-  fill: var(--co-bar-hover);
-}
-
-.catalog-overview__timeline-label {
-  fill: var(--co-fg);
-  font-size: 11px;
-  dominant-baseline: middle;
-}
-
-.catalog-overview__timeline-axis-label {
-  fill: var(--co-fg);
-  font-size: 10px;
-  opacity: 0.7;
-}
-
-.catalog-overview__timeline-no-data {
-  fill: var(--co-fg);
-  opacity: 0.5;
-  font-size: 10px;
-  font-style: italic;
-  dominant-baseline: middle;
-}
-
-/* Tooltip */
-.catalog-overview__tooltip {
-  position: fixed;
-  background: var(--co-tooltip-bg);
-  color: var(--co-tooltip-fg);
-  border: 1px solid var(--co-border);
-  padding: 4px 8px;
-  border-radius: 3px;
-  font-size: 12px;
-  pointer-events: none;
-  z-index: 10000;
-  white-space: nowrap;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.3);
-}
-
-/* Empty state */
-.catalog-overview__empty {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  opacity: 0.6;
-  font-style: italic;
-}
-
-/* Map empty state overlay — positioned inside the map region, transparent to pointer events */
-.catalog-overview__overlay {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background: rgba(0, 0, 0, 0.5);
-  color: var(--co-fg);
-  padding: 8px 16px;
-  border-radius: 4px;
-  font-size: 13px;
-  font-style: italic;
-  pointer-events: none;
-  z-index: 1000;
-  white-space: nowrap;
-}
