--- shared/components/src/CatalogOverview/index.ts
@@ -1,6 +1,0 @@
-/**
- * CatalogOverview component exports.
- */
-
-export { CatalogOverview } from './CatalogOverview';
-export type { CatalogOverviewProps, CatalogOverviewItem } from './types';
